+++ annotypes/__init__.py
@@ -1,6 +1,7 @@
 from ._anno import Anno, NO_DEFAULT
 from ._array import Array, to_array, array_type
 from ._calltypes import WithCallTypes, add_call_types, make_annotations
+from ._serializable import Serializable, serialize_object, deserialize_object
 from ._typing import (
     TYPE_CHECKING, TypeVar, Sequence, Union, Optional, Generic,
     overload, Mapping, Any, GenericMeta